Add docs and update fixtures
Added additional documentation and updated fixtures
values to accurately reflect tests. I'm noticing that
depending on the state of the database, some of these
fixture values will not satisfy the tests.

Therefore, maybe some of these values can be pulled out
into a separate static file. I don't know if that goes
against the testing philosophy which is to be upfront
about the resources a test needs.