Implement function to retrieve/store match data
The latest 100 matches can now be retrieve from the
Riot API. These matches are stored in the matchlist
table which can then be used to get match statistics.

This could be improved by allowing the user to retrieve
more than the latest 100 matches by specifying certain
parameters to enable pagination.